Fix bug with keep option.
A bug was introduced with recursive search where the tmp files would
be moved to the root of filesystem. But anyway, using rename does not
work across mount points so this option is still brittle.

We open the diff files with different filenames depending on whether
the file needs to be kept or not.

Fixes #39

+++ plotkicadsch/src/sysAbst.ml
@@ -67,22 +67,17 @@ let build_tmp_svg_name ~keep aprefix aschpath =
   let root_prefix =
     aprefix ^ String.sub aschname ~pos:0 ~len:(String.length aschname - 4)
   in
-  match detect_os () with
-  | MacOS | Linux ->
-      Stdlib.Filename.temp_file root_prefix ".svg"
-  | Cygwin | Windows ->
-      if keep then root_prefix ^ ".svg"
-      else Stdlib.Filename.temp_file root_prefix ".svg"
+  if keep then root_prefix ^ ".svg"
+  else Stdlib.Filename.temp_file root_prefix ".svg"
 
-let finalize_tmp_file fnl ~keep_as =
+let finalize_tmp_file fnl ~keep =
   match detect_os () with
   | MacOS | Linux -> (
       try%lwt
-        match keep_as with
-        | None ->
-            Lwt_unix.unlink fnl
-        | Some target ->
-            Lwt_unix.rename fnl target
+        if not keep then
+          Lwt_unix.unlink fnl
+        else
+          Lwt.return_unit
       with _ -> Lwt.return_unit )
   | Cygwin | Windows ->
       Lwt.return_unit
